Create README.md
Inception Attendance Recorder

Problem Statement:
Inception Guide is a Coaching Centre in Veraval. They face a problem in taking attendance of students and keeping a track of it. 


Proposed Solution :
This project proposes a "Attendance Recorder" to keep track of the attendance of students and information of which roll number student was absent on a particular day. It features include adding subject from the given list, add extra note for any special lecture and record attendance of students.

Functionality & Concepts used :

The App has a very simple and interactive interface which helps the faculty record attendance easily. Following are few android concepts used to achieve the functionalities in app :
Constraint Layout : Most of the activities in the app uses a flexible constraint layout, which is easy to handle for different screen sizes.
Recycler View : To present the list of different classes/lectures, we used the efficient recyclerview. 
Room Database :We have used room database to store the data of attendance locally.
The app is currently in the Alpha testing phase with Inception. 

Once the app is tested and fully functional, we plan to add some more features to it and then try to propose it for our college as well.